src/backend/requirements.txt
certifi==2019.9.11
chardet==3.0.4
Click==7.0
Flask==1.1.1
grpcio==1.23.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
protobuf==3.9.1
pycca==0.0.1
python-dotenv==0.10.3
requests==2.22.0
six==1.12.0
urllib3==1.25.5
Werkzeug==0.16.0
